feat: Move badge
